refactor(ui/components): margin components 내부로 이동
