+++ packages/ui/src/components/margin/extract-margin-props.ts
@@ -1,7 +1,7 @@
 import { assignInlineVars } from '@vanilla-extract/dynamic';
 import { type CSSProperties } from 'react';
 import * as styles from './margin.css';
-import { cx, mergeStyles, px } from '../utils';
+import { cx, mergeStyles, px } from '../../utils';
 
 export function extractMarginProps<
   P extends styles.MarginVariants & {
