+++ packages/ui/src/components/tag/tag.tsx
@@ -1,7 +1,7 @@
 import * as styles from './tag.css';
-import { type MarginVariants, extractMarginProps } from '../../margin';
 import { type HTMLFavolinkProps, favolink, forwardRef } from '../../system';
 import { cx } from '../../utils';
+import { type MarginVariants, extractMarginProps } from '../margin';
 
 export type TagProps = HTMLFavolinkProps<'span'> &
   MarginVariants &
