+++ packages/ui/src/components/typography/heading.tsx
@@ -1,8 +1,8 @@
 import * as styles from './heading.css';
 import * as commonStyles from './typography.css';
-import { type MarginVariants, extractMarginProps } from '../../margin';
 import { type HTMLFavolinkProps, Slot, forwardRef } from '../../system';
 import { cx } from '../../utils';
+import { type MarginVariants, extractMarginProps } from '../margin';
 
 export type HeadingProps = commonStyles.TypographyVariants &
   HTMLFavolinkProps<'h1'> &
